refactor: Move Obstacle rendering into obstacle.ts
