+++ imagine-sisyphus-happy/src/obstacle.ts
@@ -3,7 +3,17 @@
 // then move every obstacle down in the beat
 
 import { KeyState } from "./input";
-import { GameState, GRID_HEIGHT, GRID_WIDTH, MAX_OBSTACLES } from "./stateMachine"
+import {
+  GameState,
+  GRID_HEIGHT,
+  GRID_WIDTH,
+  MAX_OBSTACLES,
+} from "./stateMachine";
+import { Assets, Texture, Sprite, AnimatedSprite } from "pixi.js";
+
+let obstacleTexture: Texture;
+let sprite: Sprite;
+const myObstacles: Map<string, Sprite> = new Map();
 
 export type Obstacle = {
   id: string;
@@ -11,10 +21,33 @@ export type Obstacle = {
   y: number;
 };
 
+export async function initFrame(obstacles: Obstacle[]): Promise<Sprite> {
+  obstacleTexture = await Assets.load("/assets/tree.png");
+}
+
+// how am i going to keep track of myObstacles?
+// TODO: do we need to check if obstacle is already rendered?
+export async function frame(app: Application, obstacles: Obstacle[]) {
+  obstacles.forEach((obstacle) => {
+    const obstacleSprite = myObstacles.get(obstacle.id);
+    if (obstacleSprite) {
+      obstacleSprite.position.set(obstacle.x, obstacle.y);
+    } else {
+      const newSprite = new Sprite(obstacleTexture); // how to pass texture?
+      newSprite.position.set(obstacle.x, obstacle.y);
+      app.stage.addChild(newSprite);
+      myObstacles.set(obstacle.id, newSprite);
+    }
+  });
+}
+
 // mvp: generate an obstacle for every space press, and move each obstacle down by one
 // TODO: pass obstacle updates to renderer, make it it's own function?
 // TODO: handle player movement? need to calculate elevation and player position
-export function updateObstacles(obstacles: Obstacle[], expectMove: boolean): Obstacle[] {
+export function updateObstacles(
+  obstacles: Obstacle[],
+  expectMove: boolean,
+): Obstacle[] {
   // console.log("inputs: ", inputs)
   // for now, space press represents new beat
   // TODO: sync movement updates to metronome
@@ -24,17 +57,18 @@ export function updateObstacles(obstacles: Obstacle[], expectMove: boolean): Obs
     const movedObstacles = obstacles.map((obs) => ({ ...obs, y: obs.y + 20 }));
     const bounded = movedObstacles.filter((obs) => obs.y >= 0);
 
-    // TODO: possible to generate multiple obstacles per line? 
-    const spawned = bounded.length < MAX_OBSTACLES
-      ? [
-        ...bounded,
-        {
-          id: crypto.randomUUID(),
-          x: Math.floor(Math.random() * (GRID_WIDTH - 1)),
-          y: GRID_HEIGHT - 1,
-        } as Obstacle
-      ]
-      : bounded;
+    // TODO: possible to generate multiple obstacles per line?
+    const spawned =
+      bounded.length < MAX_OBSTACLES
+        ? [
+            ...bounded,
+            {
+              id: crypto.randomUUID(),
+              x: Math.floor(Math.random() * (GRID_WIDTH - 1)),
+              y: GRID_HEIGHT - 1,
+            } as Obstacle,
+          ]
+        : bounded;
 
     return spawned;
   }
