+++ imagine-sisyphus-happy/src/renderer.ts
@@ -14,13 +14,12 @@ import type { GameState } from "./stateMachine";
 import * as ui from "./ui";
 import { initDevtools } from "@pixi/devtools";
 import * as background from "./background";
-import { Obstacle } from "./obstacle";
+import * as obstacle from "./obstacle";
 import * as player from "./Player";
 import * as screen from "./backgroundScreen";
 import { DEBUG_MODE } from "./debug";
 
 let app: Application;
-let obstacleTexture: Texture;
 let lastState: GameState;
 let colorMatrix: ColorMatrixFilter;
 
@@ -29,8 +28,6 @@ export async function initialize(gameState: GameState) {
   // Create a new application
   app = new Application();
   initDevtools({ app });
-
-  obstacleTexture = await Assets.load("/assets/tree.png");
 
   await app.init({
     background: "#1099bb",
@@ -47,6 +44,9 @@ export async function initialize(gameState: GameState) {
   // TODO
   const bgScreen = screen.initFrame(app);
   app.stage.addChild(bgScreen);
+
+  const obs = await obstacle.initFrame(gameState.obstacles);
+  // app.stage.addChild(obs);
 
   const playerSprite = await player.initFrame(width, height, gameState.player);
   app.stage.addChild(playerSprite);
@@ -95,25 +95,7 @@ export async function render(state: GameState) {
 async function drawScene(state: GameState, ticker: Ticker) {
   background.frame(state, ticker);
   screen.frame(state.expectMove, app);
-  renderObstacles(app, state.obstacles);
+  obstacle.frame(app, state.obstacles);
   player.frame(state.player);
   ui.frame(state.expectMove, state.elevation, state.streak, state.lost, ticker);
-}
-
-const myObstacles: Map<string, Sprite> = new Map();
-
-// how am i going to keep track of myObstacles?
-// TODO: do we need to check if obstacle is already rendered?
-export async function renderObstacles(app: Application, obstacles: Obstacle[]) {
-  obstacles.forEach((obstacle) => {
-    const obstacleSprite = myObstacles.get(obstacle.id);
-    if (obstacleSprite) {
-      obstacleSprite.position.set(obstacle.x, obstacle.y);
-    } else {
-      const newSprite = new Sprite(obstacleTexture); // how to pass texture?
-      newSprite.position.set(obstacle.x, obstacle.y);
-      app.stage.addChild(newSprite);
-      myObstacles.set(obstacle.id, newSprite);
-    }
-  });
 }
